+++ src/Rule/Date/Date.php
@@ -130,7 +130,7 @@ final class Date extends BaseDate
         ?string $incorrectInputMessage = null,
         ?string $tooEarlyMessage = null,
         ?string $tooLateMessage = null,
-        mixed $skipOnEmpty = null,
+        bool|callable|null $skipOnEmpty = null,
         bool $skipOnError = false,
         Closure|null $when = null,
     ) {