+++ CHANGELOG.md
@@ -24,6 +24,7 @@
   (@arogachev)
 - Chg #679: Change type of `$rule` argument in `RuleHandlerInterface::validate()` from `object` to `RuleInterface`
   (@arogachev)
+- Chg #660: Change type of `$skipOnEmpty` argument in rules' constructors from `mixed` to `bool|callable|null`
 - Chg #613: Change type of `$escape` argument in `Error::getValuePath()` from `bool|string|null` to `string|null` 
   (@arogachev)